Got Display Working Correctly
I got the form and experience cards to display correctly with the right states. I now need to find a way to store the form data once submitted so I can render the correct list of cards and also be able to pull data from that object when editing a card.