Form and Resume work
I ditched the idea of making live edits to the resume as it proved to bee too complex so now the user can hit the save btn on the current experience they are editing or a new one they are adding and the changes should reflect correctly on the resume. I am working on making the Personal Details section stateful and implementing similar functionality there to the experience section.