Experience visibility and resume template work
Users can now successfully toggle the visibility of an experience on the resume by clicking the eye icon. The resume can now be toggled on and off so users can view an example and add to it or make their own from scratch. Just need to fix the issue where when user clicks cancel button on Personal Details section, it doesn't cancel because of the if statement in the PErsonalDetails component.